Add otel contextManager & use it for hub management
